+++ src/api/deleteRecordVo.ts
@@ -5,17 +5,13 @@ import type { ClientConfiguration } from '../types';
 export const deleteRecordVo = async (
   clientConfiguration: ClientConfiguration,
   recordId: number,
-): Promise<boolean> => {
+): Promise<void> => {
   const queryParams = new URLSearchParams({
     recordId: `${recordId}`,
   });
-  const response = await makePermanentApiCall(
+  await makePermanentApiCall(
     clientConfiguration,
     `/record/delete?${queryParams.toString()}`,
     { method: 'DELETE' },
   );
-  if (response.status === 200) {
-    return true;
-  }
-  return false;
 };