+++ src/errors/index.ts
@@ -1,1 +1,2 @@
+export * from './HttpResponseError';
 export * from './ValidationError';