+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- This package no longer supports Node v14
+## Changed
+- This package no longer supports Node v14.
+- The `deleteFolder` and `deleteArchiveRecord` SDK methods now return void (was boolean).
+- All SDK methods now throw if the server responds with a non-OK status.
+
+## Added
+- The `HttpResponseError` type now exists.
 
 ## [0.6.0] - 2023-05-18
 ## Changed